use embedded structs
